refactor(crypto): move crypto adapter to services directory and add comprehensive tests
