+++ src/infrastructure/services/crypto/__test__/crypto.service.test.ts
@@ -1,4 +1,4 @@
-import { CryptoAdapter } from "@/infrastructure/adapters/crypto.adapter";
+import { CryptoAdapter } from "@/infrastructure/services/crypto/adapter/crypto.adapter";
 import { CryptoService } from "@/infrastructure/services/crypto/crypto.service";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
