+++ src/infrastructure/services/crypto/adapter/crypto.adapter.ts
